tools/requirements.txt
shapely>=2.0
pyyaml>=6.0
